nst_/requirements.txt
certifi==2021.10.8
cffi @ file:///opt/concourse/worker/volumes/live/506ff722-306a-4142-6379-cfea99aadbed/volume/cffi_1625814701363/work
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
numpy @ file:///opt/concourse/worker/volumes/live/5572694e-967a-4c0c-52cf-b53d43e72de9/volume/numpy_and_numpy_base_1603491881791/work
olefile @ file:///Users/ktietz/demo/mc3/conda-bld/olefile_1629805411829/work
Pillow @ file:///opt/concourse/worker/volumes/live/a7c937c5-8583-47b2-5ff1-0dd2a9a0d739/volume/pillow_1625655830909/work
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
six @ file:///tmp/build/80754af9/six_1623709665295/work
torch @ file:///opt/concourse/worker/volumes/live/6dab859c-71b2-4f03-64f1-a28f9654f551/volume/pytorch_1614406971210/work
torchvision @ file:///opt/concourse/worker/volumes/live/6d58f683-7a53-469e-6315-7367e04ea7a9/volume/torchvision_1622219699179/work
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
